arrumado erro q dava por causa da diferença entre double e int do atributo saldo da conta

+++ src/Views/TelaDOC.java
@@ -183,7 +183,7 @@ public class TelaDOC extends javax.swing.JFrame {
         //
         //
         if (tela.confirma()){
-            int saldo =usuario.getConta().getSaldo();
+            double saldo =usuario.getConta().getSaldo();
             int valor=Integer.parseInt(campoValor.getText());
             if (saldo >= valor){
                 if (valor < 5000){
